tests/wpt/web-platform-tests/app-history/navigate/navigate-state-repeated.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  appHistory.onnavigate = e => e.transitionWhile(Promise.resolve());

  let p1 = promise_rejects_dom(t, 'AbortError', appHistory.navigate('/foo', {state: {foo: 1}}));
  assert_equals(appHistory.current.getState().foo, 1);
  let p2 = appHistory.navigate('/foo', {state: {foo: 2}});
  assert_equals(appHistory.current.getState().foo, 2);
  await p1;
  await p2;
}, "navigate() with state should work correctly when called repeatedly");
</script>
